Move ConfirmPaymentIntentParamsFactoryTest
