+++ payments-core/src/test/java/com/stripe/android/ConfirmPaymentIntentParamsFactoryTest.kt
@@ -1,31 +1,26 @@
-package com.stripe.android.paymentsheet.model
+package com.stripe.android
 
-import com.google.common.truth.Truth
-import com.stripe.android.model.CardBrand
+import com.google.common.truth.Truth.assertThat
+import com.stripe.android.model.Address
 import com.stripe.android.model.ConfirmPaymentIntentParams
 import com.stripe.android.model.PaymentMethodCreateParamsFixtures
 import com.stripe.android.model.PaymentMethodFixtures
 import com.stripe.android.model.PaymentMethodOptionsParams
-import com.stripe.android.paymentsheet.PaymentSheet
-import com.stripe.android.paymentsheet.addresselement.AddressDetails
-import com.stripe.android.paymentsheet.addresselement.toConfirmPaymentIntentShipping
 import org.junit.Test
 
 class ConfirmPaymentIntentParamsFactoryTest {
+
     private val factory = ConfirmPaymentIntentParamsFactory(
-        PaymentIntentClientSecret(CLIENT_SECRET),
-        null
+        clientSecret = CLIENT_SECRET,
+        shipping = null,
     )
 
     @Test
     fun `create() with new card when savePaymentMethod is true should create params with setupFutureUsage = OffSession`() {
-        Truth.assertThat(
+        assertThat(
             factory.create(
-                paymentSelection = PaymentSelection.New.Card(
-                    PaymentMethodCreateParamsFixtures.DEFAULT_CARD,
-                    CardBrand.Visa,
-                    customerRequestedSave = PaymentSelection.CustomerRequestedSave.RequestReuse
-                )
+                createParams = PaymentMethodCreateParamsFixtures.DEFAULT_CARD,
+                setupFutureUsage = ConfirmPaymentIntentParams.SetupFutureUsage.OffSession,
             )
         ).isEqualTo(
             ConfirmPaymentIntentParams.createWithPaymentMethodCreateParams(
@@ -41,13 +36,10 @@ class ConfirmPaymentIntentParamsFactoryTest {
 
     @Test
     fun `create() with new card when savePaymentMethod is true should create params with setupFutureUsage = blank`() {
-        Truth.assertThat(
+        assertThat(
             factory.create(
-                paymentSelection = PaymentSelection.New.Card(
-                    PaymentMethodCreateParamsFixtures.DEFAULT_CARD,
-                    CardBrand.Visa,
-                    customerRequestedSave = PaymentSelection.CustomerRequestedSave.RequestNoReuse
-                )
+                createParams = PaymentMethodCreateParamsFixtures.DEFAULT_CARD,
+                setupFutureUsage = ConfirmPaymentIntentParams.SetupFutureUsage.Blank,
             )
         ).isEqualTo(
             ConfirmPaymentIntentParams.createWithPaymentMethodCreateParams(
@@ -63,13 +55,10 @@ class ConfirmPaymentIntentParamsFactoryTest {
 
     @Test
     fun `create() with new card when savePaymentMethod is true should create params with setupFutureUsage = null`() {
-        Truth.assertThat(
+        assertThat(
             factory.create(
-                paymentSelection = PaymentSelection.New.Card(
-                    PaymentMethodCreateParamsFixtures.DEFAULT_CARD,
-                    CardBrand.Visa,
-                    customerRequestedSave = PaymentSelection.CustomerRequestedSave.NoRequest
-                )
+                createParams = PaymentMethodCreateParamsFixtures.DEFAULT_CARD,
+                setupFutureUsage = null,
             )
         ).isEqualTo(
             ConfirmPaymentIntentParams.createWithPaymentMethodCreateParams(
@@ -83,59 +72,49 @@ class ConfirmPaymentIntentParamsFactoryTest {
 
     @Test
     fun `create() with saved card and shippingDetails sets shipping field`() {
-        val shippingDetails = AddressDetails(
+        val shippingDetails = ConfirmPaymentIntentParams.Shipping(
             name = "Test",
-            address = PaymentSheet.Address(
+            address = Address(
                 line1 = "line1",
-                city = "city"
+                city = "city",
             ),
-            phoneNumber = "5555555555"
-        )
-        val factoryWithConfig = ConfirmPaymentIntentParamsFactory(
-            PaymentIntentClientSecret(CLIENT_SECRET),
-            shippingDetails.toConfirmPaymentIntentShipping()
+            phone = "5555555555",
         )
 
-        Truth.assertThat(
-            factoryWithConfig.create(
-                paymentSelection = PaymentSelection.Saved(
-                    PaymentMethodFixtures.CARD_PAYMENT_METHOD
-                )
-            ).shipping
-        ).isEqualTo(
-            shippingDetails.toConfirmPaymentIntentShipping()
+        val factoryWithConfig = ConfirmPaymentIntentParamsFactory(
+            clientSecret = CLIENT_SECRET,
+            shipping = shippingDetails,
         )
+
+        val result = factoryWithConfig.create(PaymentMethodFixtures.CARD_PAYMENT_METHOD)
+        assertThat(result.shipping).isEqualTo(shippingDetails)
     }
 
     @Test
     fun `create() with new card and shippingDetails sets shipping field`() {
-        val shippingDetails = AddressDetails(
+        val shippingDetails = ConfirmPaymentIntentParams.Shipping(
             name = "Test",
-            address = PaymentSheet.Address(
+            address = Address(
                 line1 = "line1",
-                city = "city"
+                city = "city",
             ),
-            phoneNumber = "5555555555"
-        )
-        val factoryWithConfig = ConfirmPaymentIntentParamsFactory(
-            PaymentIntentClientSecret(CLIENT_SECRET),
-            shippingDetails.toConfirmPaymentIntentShipping()
+            phone = "5555555555",
         )
 
-        Truth.assertThat(
-            factoryWithConfig.create(
-                paymentSelection = PaymentSelection.New.Card(
-                    PaymentMethodCreateParamsFixtures.DEFAULT_CARD,
-                    CardBrand.Visa,
-                    customerRequestedSave = PaymentSelection.CustomerRequestedSave.NoRequest
-                )
-            ).shipping
-        ).isEqualTo(
-            shippingDetails.toConfirmPaymentIntentShipping()
+        val factoryWithConfig = ConfirmPaymentIntentParamsFactory(
+            clientSecret = CLIENT_SECRET,
+            shipping = shippingDetails,
         )
+
+        val result = factoryWithConfig.create(
+            createParams = PaymentMethodCreateParamsFixtures.DEFAULT_CARD,
+            setupFutureUsage = null,
+        )
+
+        assertThat(result.shipping).isEqualTo(shippingDetails)
     }
 
     private companion object {
-        private const val CLIENT_SECRET = com.stripe.android.paymentsheet.PaymentSheetFixtures.CLIENT_SECRET
+        private const val CLIENT_SECRET = "client_secret"
     }
 }
